// src/app/mill_management_system/model/Masters/mill-master.model.ts
export class MillMaster {
    millName: string | any;
    stateCode: string | any;
    districtCode: string | any;
    talukaCode: string | any;
    millLocation: string | any;
    contactPerson: string | any;
    contactNo: string | any;
    address: string | any;
    emailId:String|any;
    userId:String|any;
      
}